+++ .github/CHANGELOG.md
@@ -18,4 +18,4 @@
 17. [Game] Test - Added tests to check UIUtils class
 18. [Game] Enhancement - Increased default window size to 1440x720 and changed map zoom factor accordingly
 19. [Global] Enhancement - Added actions to release snapshot and stable versions
-
+20. [Game] Enhancement - Added achievements system